Refine indexation

+++ lib/zek/cmd_delete.rb
@@ -9,7 +9,11 @@ module Zek::CmdDelete; class << self
 
     Zek.note_files(Zek.lookup_uuid(args.first)).each do |path|
 
-      FileUtils.rm_f(path)
+      if File.extname(path) == '.md'
+        File.open(path, 'ab') { |f| f.write("\n<!-- status: deleted -->\n") }
+      else
+        FileUtils.rm_f(path)
+      end
     end
   end
